+++ Adapter/ShopwareAdapter/DataProvider/Translation/TranslationDataProviderInterface.php
@@ -4,6 +4,7 @@ namespace ShopwareAdapter\DataProvider\Translation;
 
 use Shopware\Models\Article\Configurator\Group as ConfiguratorGroupModel;
 use Shopware\Models\Article\Configurator\Option as ConfiguratorOptionModel;
+use Shopware\Models\Article\Image;
 use Shopware\Models\Property\Option as OptionModel;
 use Shopware\Models\Property\Value as ValueModel;
 use Shopware\Models\Shop\Shop as ShopModel;
@@ -47,4 +48,12 @@ interface TranslationDataProviderInterface
      * @return null|ConfiguratorOptionModel
      */
     public function getConfigurationOptionByName(Value $value);
+
+    /**
+     * @param Identity $mediaIdentity
+     * @param $articleId
+     *
+     * @return Image
+     */
+    public function getArticleImage(Identity $mediaIdentity, $articleId);
 }