+++ Adapter/ShopwareAdapter/ServiceBus/CommandHandler/Product/HandleProductCommandHandler.php
@@ -147,6 +147,12 @@ class HandleProductCommandHandler implements CommandHandlerInterface
             } else {
                 $this->correctMainDetailAssignment($mainVariation);
 
+                foreach ($mainVariation->getImages() as $image) {
+                    if (null !== $image) {
+                        $this->translationDataPersister->removeMediaTranslation($image);
+                    }
+                }
+
                 $productModel = $articleResource->update($mainVariation->getArticleId(), $params);
             }
 
@@ -170,6 +176,12 @@ class HandleProductCommandHandler implements CommandHandlerInterface
                         'number' => $product->getNumber(),
                         'active' => true,
                     ]);
+                }
+
+                foreach ($productModel->getImages() as $image) {
+                    if (null !== $image) {
+                        $this->translationDataPersister->removeMediaTranslation($image);
+                    }
                 }
 
                 $this->correctMainDetailAssignment($mainVariation);