+++ Connector/TransferObject/Product/Badge/Badge.php
@@ -46,7 +46,6 @@ class Badge extends AbstractValueObject
     public function getClassProperties()
     {
         return [
-            'class' => self::class,
             'type' => $this->getType(),
         ];
     }