+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[## [unreleased]]
+### Fixed
+- fix category sync if more then one shop has the same root category (@lacodimizer)
+
 ## [## [5.1.0]]
 ### Fixed
 - transfer payment information without a real transactionid if needed