+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[## [unreleased]]
+## [## [unreleased]]
 ### Fixed
 - fix category sync if more then one shop has the same root category (@lacodimizer)