+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - article image translation (@Pfabeck)
+- amazonPay by BestIT integration (@Pfabeck)
 
 ## [## [5.1.0]]
 ### Fixed